unseal_service/requirements.txt
djangorestframework==3.16.0
Django==5.2.3
cryptography==45.0.4
